*** empty log message ***

+++ trunk/IPK-Plugins/src/java/de/ipk_gatersleben/ag_nw/graffiti/services/task/ProgressStatusService.java
@@ -87,28 +87,28 @@ public class ProgressStatusService implements HelperClass {
 		int nn = 1;
 		String res[] = new String[2];
 		if (remainTimeYears > 1 && numberResults < nn)
-			if (remainTimeYears == 1)
+			if (((int)remainTimeYears) == 1)
 				res[numberResults++] = (int) remainTimeYears + "&nbsp;year";
 			else
 				res[numberResults++] = (int) remainTimeYears + "&nbsp;years";
 		// if (remainTimeMonths>1 && numberResults<2)
 		// res[numberResults++]=(int)remainTimeMonths+" months";
 		if (remainTimeWeeks > 1 && numberResults < nn)
-			if (remainTimeWeeks == 1)
+			if (((int)remainTimeWeeks) == 1)
 				res[numberResults++] = (int) remainTimeWeeks + "&nbsp;week";
 			else
 				res[numberResults++] = (int) remainTimeWeeks + "&nbsp;weeks";
-		if (remainTimeDays > 1 && numberResults < nn)
+		if (((int)remainTimeDays) > 1 && numberResults < nn)
 			if (remainTimeDays == 1)
 				res[numberResults++] = (int) remainTimeDays + "&nbsp;day";
 			else
 				res[numberResults++] = (int) remainTimeDays + "&nbsp;days";
-		if (remainTimeHours > 1 && numberResults < nn)
+		if (((int)remainTimeHours) > 1 && numberResults < nn)
 			if (remainTimeHours == 1)
 				res[numberResults++] = (int) remainTimeHours + "&nbsp;hour";
 			else
 				res[numberResults++] = (int) remainTimeHours + "&nbsp;hours";
-		if (remainTimeMinutes > 1 && numberResults < nn)
+		if (((int)remainTimeMinutes) > 1 && numberResults < nn)
 			res[numberResults++] = (int) remainTimeMinutes + "&nbsp;min";
 		if (numberResults < nn)
 			res[numberResults++] = (int) remainTimeSeconds % 60 + "&nbsp;sec";
